+++ login-workflow/src/guards/auth-guard.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import { Router, CanActivate } from '@angular/router';
 import { PxbAuthSecurityService } from '../services/state/auth-security.service';
 import { PxbAuthConfig } from '../services/config/auth-config';
+import { AUTH_ROUTES } from '../auth/auth.routes';
 
 @Injectable({
     providedIn: 'root',
@@ -15,7 +16,11 @@ export class PxbAuthGuard implements CanActivate {
 
     canActivate(): boolean {
         if (!this.securityService.getSecurityState().isAuthenticatedUser) {
-            void this.router.navigate([this.pxbAuthConfig.authGuardRedirectRoute]);
+            const route =
+                this.pxbAuthConfig.authGuardRedirectRoute === undefined
+                    ? AUTH_ROUTES.AUTH_WORKFLOW
+                    : this.pxbAuthConfig.authGuardRedirectRoute;
+            void this.router.navigate([route]);
             return false;
         }
         return true;