+++ login-workflow/src/pages/create-account-invite/create-account-invite.component.ts
@@ -31,6 +31,7 @@ export class PxbCreateAccountInviteComponent implements OnInit, OnDestroy {
 
     isLoading: boolean;
     isValidRegistrationLink: boolean;
+    isPXWhiteAccount: boolean;
 
     // EULA Page
     userAcceptsEula: boolean;
@@ -78,9 +79,10 @@ export class PxbCreateAccountInviteComponent implements OnInit, OnDestroy {
         this._pxbSecurityService.setLoading(true);
         this._pxbRegisterService
             .validateUserRegistrationRequest()
-            .then(() => {
+            .then((registrationComplete) => {
                 this._pxbSecurityService.setLoading(false);
                 this.isValidRegistrationLink = true;
+                this.isPXWhiteAccount = registrationComplete;
             })
             .catch((data: ErrorDialogData) => {
                 this._pxbErrorDialogService.openDialog(data);