+++ login-workflow/src/translations/english.ts
@@ -125,5 +125,10 @@ export const pxbAuthEnglishTranslations: PxbAuthTranslations = {
             },
             CONTINUE_BUTTON: 'Continue',
         },
+        ACCOUNT_EXISTING: {
+            TITLE: 'Account Created',
+            WELCOME_MESSAGE_DESCRIPTION: 'Your account has been successfully created. Please log in with your Eaton account email and password.',
+            WELCOME_MESSAGE_TITLE: 'Welcome'
+        }
     },
 };