+++ login-workflow/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v2.1.0 (not published)
 
 -  Added additional `pageTitle` and `pageInstructions` props onto `AccountDetails` for more customization during account registration.  
+-  Added `registrationSuccessScreen` prop to `<pxb-create-account>` & `<pxb-create-account-invite>` component for success screen customization.
 
 ## v2.0.0 (March 31, 2021)