+++ login-workflow/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 -  Added additional `pageTitle` and `pageInstructions` props onto `AccountDetails` for more customization during account registration.  
 -  Added `registrationSuccessScreen` prop to `<pxb-create-account>` & `<pxb-create-account-invite>` component for success screen customization.
+-  Added `authGuardRedirectRoute` prop to `PxbAuthConfig`. 
 
 ## v2.0.0 (March 31, 2021)